Unify GLInstances and Oauth mechanism
This commit updates the GLInstances and now are used in the same way.
UI part is missing but that will come in a future commit.

GitLab Oauth mechanism is now unified in one to make it simpler to include
future backends.

Kyes for GitLab instances are read from settings file and included in the database in run-time

Signed-off-by: JJMerchante <jj.merchante@gmail.com>